re add missing prettier add new folder structure
