+++ package.json
@@ -32,7 +32,8 @@
     "@typescript-eslint/parser": "^4.33.0",
     "electron": "^24.3.0",
     "eslint": "^7.32.0",
-    "typescript": "^4.8.3"
+    "typescript": "^4.8.3",
+    "prettier": "^3.0.0"
   },
   "dependencies": {
     "electron-squirrel-startup": "^1.0.0"
